kafka-consumer/requirements.txt
kafka-python==2.0.2
pytest==7.4.0